+++ packages/illustrations/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.5.1](https://github.com/Gympass/yoga/compare/@gympass/yoga-illustrations@0.5.0...@gympass/yoga-illustrations@0.5.1) (2022-09-20)
+
+**Note:** Version bump only for package @gympass/yoga-illustrations
+
+
+
+
+
 # [0.5.0](https://github.com/Gympass/yoga/compare/@gympass/yoga-illustrations@0.4.0...@gympass/yoga-illustrations@0.5.0) (2022-07-28)